refactor: add .web to desktop components
