+++ devtools/shell/src/components/IconSquare.web.tsx
