+++ devtools/shell/src/components/SectionHeader.web.tsx
