+++ devtools/shell/src/components/index.ts
@@ -1,9 +1,9 @@
-export { ToolRow } from "./ToolRow";
-export { CategoryRow } from "./CategoryRow";
-export { CategoryCard } from "./CategoryCard";
-export { ToolCard } from "./ToolCard";
-export { IconSquare } from "./IconSquare";
-export { SectionHeader } from "./SectionHeader";
-export { Sidebar } from "./Sidebar";
-export { ToolShell } from "./ToolShell";
-export { Overview } from "./Overview";
+export { ToolRow } from "./ToolRow.web";
+export { CategoryRow } from "./CategoryRow.web";
+export { CategoryCard } from "./CategoryCard.web";
+export { ToolCard } from "./ToolCard.web";
+export { IconSquare } from "./IconSquare.web";
+export { SectionHeader } from "./SectionHeader.web";
+export { Sidebar } from "./Sidebar.web";
+export { ToolShell } from "./ToolShell.web";
+export { Overview } from "./Overview.web";
